standalone_testers/requirements.txt
pycryptodome==3.17
Requests==2.31.0
